Jesus fix distance accuracy

+++ server/src/main/java/com/tco/misc/DistanceCalculator.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public class DistanceCalculator {
 
-    public Integer calculateGreatCircleDistance(Map<String,String> place1, Map<String,String> place2, Float radius)
+    public Long calculateGreatCircleDistance(Map<String,String> place1, Map<String,String> place2, Double radius)
     {
         Double long1 = Math.toRadians(Double.parseDouble(place1.get("longitude")));
         Double lat1 = Math.toRadians(Double.parseDouble(place1.get("latitude")));
@@ -34,7 +34,7 @@ public class DistanceCalculator {
         // arctan2(y/x)
         Double centralAngle = radius * (Math.atan2(numerator,denominator));
 
-        Integer distance = (int) Math.round(centralAngle);
+        Long distance = Math.round(centralAngle);
 
         return  distance;
     }
